annoncewithuserdone

+++ app/Http/Controllers/AnnonceController.php
@@ -3,7 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\annonce;
+use App\Models\User;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class AnnonceController extends Controller
 {
@@ -12,8 +15,10 @@ class AnnonceController extends Controller
      */
     public function index()
     {
-        $annonces = Annonce::latest()->where('taken', false)->filter(request(['tag' , 'search']))->get(); 
-    
+        $annonces = Annonce::latest()->where('taken', false)->filter(request(['tag' , 'search']))->with(['user' => function ($query) {
+            $query->select('id', 'username','phone', 'profile_photo_path');
+        }])->get(); 
+
         return view('Userinterface.Annonce.index', [
             'annonces' => $annonces ,
         ]);
@@ -46,7 +51,8 @@ class AnnonceController extends Controller
         if($request->Hasfile('imageannonce')){
             $formFields['image']=$request->file('imageannonce')->store('imageannonces','public');
         }
- 
+   
+        $formFields['user_id'] = auth()->id(); 
         annonce::create($formFields);
         
         return redirect()->route('Annonce.index')
@@ -58,8 +64,12 @@ class AnnonceController extends Controller
      */
     public function show(string $id)
  
-    {    
+    {   
+     
         $annonce = annonce::with('exchangesdemands')->find($id);
+        if (Gate::denies('manage', $annonce)) {
+            abort(403, 'Unauthorized action.'); 
+        } 
         return view('Userinterface.Annonce.show', compact('annonce')) ;
     }
 
@@ -69,6 +79,10 @@ class AnnonceController extends Controller
     public function edit(string $id)
     {  
         $annonce = annonce::find($id) ;
+
+        if (Gate::denies('manage', $annonce)) {
+            abort(403, 'Unauthorized action.'); 
+        }
        return view('Userinterface.Annonce.edit', [
         'annonce' => $annonce
     ]) ;
@@ -78,8 +92,11 @@ class AnnonceController extends Controller
      * Update the specified resource in storage.
      */
     public function update(Request $request, string $id)
-    {
+    {  
+
+      
         $annonce = annonce::find($id) ;
+    
         $formFields = $request->validate(
             [
                 'title' => 'required',
@@ -107,6 +124,9 @@ class AnnonceController extends Controller
     {   
      
         $annonce = annonce::find($id) ;
+        if (Gate::denies('manage', $annonce)) {
+            abort(403, 'Unauthorized action.'); 
+        } 
         if($annonce){
             $annonce->delete() ;
         }
@@ -116,8 +136,12 @@ class AnnonceController extends Controller
     }
 
     public function UserList()
-    {   
-        $annonces = Annonce::latest()->get(); 
+    {
+   
+        $user = User::find(auth()->id());
+        
+     
+        $annonces = $user->annonces()->latest()->get();
     
         return view('components.profile', [
             'annonces' => $annonces
